I got up early and for once I feel like starting right away. Let me just shake my grogginess. I'll get into it as I go along.
...Yeah, it is difficult to just jump into it. I've been staring at this commit for 10m now.

Once I am done with join points and layout types things will get easier. And I am mostly done with join points.

I swear all this work would be absolutely unbearable if not for the fact that every time I start I get into it.

It is ironic, but easy and hard tasks are both equaly difficult for me. Easy stuff is harder than it should be as I have to fight boredom every step of the way, and hard stuff is easier than it should be as I slip into the zone.

...

Agh, I tried. Let me just get an apple or something. I'll start in 15-30m when life comes into my body.