"11:20am. Let me resume.
11:30am.

```
            match t with
            | LayoutHeapMutable -> push_op_no_rewrite d LayoutToNone v (type_get data)
            | _ -> push_op d LayoutToNone v (type_get data)
```

11:35am. I am just thinking and thinking and thinking - is there anything I can do speed the language past this point?

I think I am going to dispense with functions in modules. Forget about `open`ing them. For the sake of fast compilation I will make them all objects. Modules like `Struct` and `Tuple` should never need to be opened. The only thing which ever really needs that sort of thing would be `Core`.

The memoization that I am doing right now is great, but would it be possible to go even higher than now?

11:55am. Yeah, I love the layout types, but unboxing them on every access is .... mmmmmhhhh.

12:10pm. I am having a surge of inspiration.

That is right - there is no need to ever unbox layout types for the sake of application. There would be no issue at all to just grab the environment inside the layout type and partially evaluate the relevant function or object. I just need to give the code generator some indication that it should rename the variables in the global environment to the layout type accesses.

12:15pm. That would be easy. Unlike the partial evaluator the code generator needs to be aware of the scope it is in.

This would save me such a massive amount of work in join points. It would speed up the language tremendously.

12:20pm. Ah, damn I did something really stupid.

```
and [<CustomComparison;CustomEquality>] TyTag =
    | T of Tag * ConsedTy

    override a.Equals(b) =
        match b with
        | :? TyTag as b -> match a,b with T(a,_), T(b,_) -> a = b
        | _ -> false
    override a.GetHashCode() = match a with T(a,_) -> a
    interface IComparable with
        member a.CompareTo(b) =
            match b with
            | :? TyTag as b -> match a,b with T(a,_), T(b,_) -> compare a b
            | _ -> raise <| ArgumentException "Invalid comparison for TyTag."
```

So an int and a float are the same as long as they have the same tag? Ah crap.

12:25pm. Removed. I can definitely push work to the code generator without impacting compilation speed much at all.

I'll continue this line of thought later.

Right now I am overdue for breakfast."