Yesterday's session hit me like a bag of bricks when it was done. I am not sure what drug withdrawal is like since I never tried cocaine, but it surely must be like that. It was a deep fatigue that crushed me from the inside, so heavy that the only way for me to sleep is to fall into unconsciousness due to strain.
But still, I want to keep going despite the residue of it still lingering. While I am alive I want to keep going for as long as I can. v0.1 of Spiral is worth it. Despite some of its shortfalls Spiral is still the most powerful and expressive statically typed language in the world.

I am not sure whether once I am done with the new language I will feel like rewriting ALL the libraries. Poker with ML is no longer a priority for me, but I want to get some decent testing done. I'll always be able to get the Cuda stuff back online once it is needed.

Because of the scale of the changes what I am doing is essentially rebuilding the language from the ground up and given that at this point it is not a short project, I can't just dump two months of my life whenever I feel like making a change. I think I've already been working for 2 weeks on this.

I should just do what I want.

I do not have any users or bosses standing over my shoulder and I have nothing to really prove. And I do not need to be held accountable to my previous work. The stuff that I did in v0.09 is proof enough of my skill.

Cuda and deep learning RL had a chance - if they really were useful then there would be no need to hesitate on getting them back online, but they are stupidly useless.

But Spiral is worth it. Maybe the current wave of ML will die, or maybe backpropagation will evolve into the metaheuristic algorithm somehow. I have no idea what will happen in the end with backpropagation's story.

But ultimately, my vision of a need for a language with extremely high performance for the widest possible range of domains and devices will come to bear. Languages like Julia are just entertainment for the Matlab crowd, and systems languages like Rust do not have the power nor expressiveness to compete with Spiral.

v0.1 of Spiral will be the first (half-way) professionaly done Spiral. When I finished Spiral a year ago I was satisfied with the design, but deep down I've been hoping I would get implementation advice from those more experienced than me as I knew Spiral was lacking in areas. There were issues that I had no idea how to deal with.

I no longer need such advice. I KNOW how to implement a language such as Spiral properly now.

And that is what this will be about. A year after its release, I am improving the language based on my experience.

After I am done, there will still be a lot of overhead left in the parser, but that never worried me. Unlike the partial evaluator I always knew how to speed that up - more inlining.

Spiral is the Earth and Pharo will be the Heaven. The way I've been programming in Spiral for the past year is a good way to fail at all my goals, but I can do so much better with more effort.

The way I've been doing ML without any visualization and with only a very sparse input is quite reminiscent of how I used to trade by just watching the price move. Unlike in trading, I have access to so much more informantion about the system and yet I am throwing it all away. That is a good way to carve my failure in stone.

ML is not the kind of thing where logic will see me through - I need to find a way to plug myself into the process.

I will absolutely never find the metaheuristic algorithm if I just keep doign things at random like the rest of the field.

A true programmer would not do it like this.

A true programmer would find a way to fully tap into the power of the machines.

Having the capability to write efficient DSLs is just a part of it. Having a high enough capacity for virtualization is also necessary for the sake of writing proper interfaces to them.

That is what the next few years are going to be all about.

I am going to move to the interface part without the agents and see where that gets me.

It is too bad that I do not have the algorithm that I wanted, but I know that for poker specifically I can always just use CFR and if it can be run on a supercomputer, maybe I could also find a way to port it to the GPU.

I've been able to meet my expressiveness goals with Spiral, but I am still yet to experience that feeling of making a definite gain on a subject of interest.

Porting CFR to the GPU using Spiral seems like it could be the ticket.

...Now, let me finally start for the day. I think I will just do some light programming in the next half hour, maybe deal with the `Record` cases, and stop for breakfast.

I should have started an hour ago, but I am still heavily reeling from yesterday and this rant is a good way to waste time in a productive manner. By reaffirming my will, I build my motivation for the tasks that are to come.

Let me do it.