I've been chilling for nearly an hour now so I am ready to start. It is time to get into it.
I'll stop looking down on the codegen and give it the attention it properly deserves. Is there any point in saying this or that is easy?

Unlike the partial evaluator...or rather exactly like it, codegen in fact does need a redesign from the ground up. I completely changed the semantics of how destructuring is done and things will have to be changed to reflect that.

Also, there are some things I dislike from last time like that buffer swapping thing.

My sense is that the overall structure of the codegen is too loose. The thing needs a redesign.

It is kind of difficult.

It might be more difficult than I think.

But let me leave behind my prejudices and do it as I desire. I will manifest the weight of my experience in making this and see where that takes me.

Forget what was before. Let me start by sketching out what I think I need. I should be able to start with the environment's type.