"1:45pm. Done with chores. Now I am free.
Let me get into it. Right now it is annoying. I'd rather read The Gamer instead of program, but two hours ago I was so into it.

Well, I'll get into it again.

1:50pm. Ok. Before I stopped last time I've been thinking of pivoting in a completely other direction. Since I am working here to speed things up, I might as well speed up everything to its limit and that includes the code generator.

Since at this point all the `Op`s have been linearized, I can in fact make a whole sweep over the method and do the string building that way.

There is no need to ever return anything.

I should speed up string construction as much as possible.

I am not a beginner here anymore. I know how to handle the flow.

```
type ProgramNode =
    | Text of string
    | Statement of string
    | Statements of string []
    | Indent
    | Dedent
```

This is going right out.

```
type CodegenEnv() =
    let types_printed = HashSet(HashIdentity.Structural)
    let types_unprinted = HashSet(HashIdentity.Structural)
    let join_points_printed = HashSet(HashIdentity.Structural)
    let join_points_unprinted = HashSet(HashIdentity.Structural)
    let buffer = ResizeArray()
```

Also I will get this out.

```
type CodegenEnv =
    {
    stmts : StringBuilder
    ret : TypedData option
    }
```

This should be the environment.

I do not want to fill it with methods, it is fine if I do it functionally.

It is just - how do I keep track of which types and methods need to be printed. This was a pain in the ass last time as well.

You know what, I will just ignore the issue of that. Once I am done it will just be a matter of adding the relevant `HashSet`s in order to keep track of that.

I do not want to wrack my mind over that right now.

Let me get the whole codegen done and then I will deal with this when it comes to print the type definitions."