Let me start. I realized that I forgot StringSlice during the night so let me start with that.
I do not think I'll be able to finish the codegen today, but I should be able to go through most of it.

Let me put in a decent effort. It will be such a relief when I am finally done with all of this and am able to move to testing.