I am ready to resume. So what is next?
The printing of methods and type definitinos.

I have difficulty in words explaining how just good of a job I did with the language so far.

I think that in the module prepass I could go a bit further in simplifying it by removing the `open` there, but...

I'll consider it when this is done.

There is no real need to do that as it would just make compilation slower. It is fine.

...

Right now I am dwelling on the future. The time will soon arrive when things will start to get serious.

I will move from the command line and to graphical interfaces. It will be a marked jump in sophistication to how used to do programming before. All the bountiful talent that I had and all the hard work had not been able to save me from failure.

The story of constant improvement in supervised learning that I tried to reenact in reinforcement learning could not get off the ground.

Once more, grind my programming into pieces and bring them back together again. I will pursue a meta-strategy of absolute integration and understanding from here on out. I will not make anymore moonshot bets. I will not bear the humiliation of not understanding what I am doing ever again.

If I try ML again, it will be with the assurance that I have the tools necessary to handle the learning process properly. No more just tracking the cost or winrates.

I will be looking at the what the process truly is.

I will seek out a small victory in counterfactual regret minimization and then build up on it. I will replace text with images and make the world come alive. My plan is perfect - ordinarily I would not place my fate in a slow dynamic language such as Pharo, but Spiral is the answer to all performance woes anywhere. The potential is there to make an absolute step upwards in productivity - I just need to invest the time necessary to make it a reality.

Already I am tired, but I must go on.

Types and methods - let me do them.