I got up 30m ago, but I do not not feel like starting just yet.
I worked so hard over the last two weeks, it would be insane to keep this pace up all the time.

...Let me rest for another half an hour so I can shake of my grogginess and then I'll think about what comes next.

Today I will get the testing scaffolding back online and maybe do the starter tests. I'll see how it goes.

Once thing I should do is start work on the changelog for v0.1. It is best that I start early as I do the tests so that I do not forget the particular features that have been done.

Starting out with a summary for the changelog seems like a decent way to start the thing. I'll do it after I put in the testing harness.