Yawn. Well, it is time to start this thing.
I'll have to go through all these tests by hand one way or another. Well, it should go quickly once I get into it. The hardest parts will be at the beginning.

The sooner I start, the sooner I finish. After this is done I really will leave Spiral alone for a while and do some programming in Pharo as I had enough of it.

I want to be free from this endless self impossed drudgery. Yes, it is amazing how well it came out and it is worth it just for that, but I want more.

If only I could program much faster. Much better. That is what I desire now.

ML is something that should be mastered by starting from the Bayesian perspective and baking in various optimizations starting from there.

Deep learning is just Bayesian inference + a very specific set of optimizations made to arrive at that point. Of course at the moment what those optimizations are is unknown. And as the assumptions behind them are unknown as well, that means I am completely at a loss of how to adapt the algorithms so they work for RL.

In the future, without a doubt there will be more principled ways of doing learning using machines.

And if I play my cards right I might find them myself.

So that I might start the search, I need to put Spiral in a box so to speak and make it plug and play for other languages.

That will allow me to declare a job well done and move on from it to bigger and better things.

I want to do things of importance. I've given up on ML for the time being, but I want to keep the quest to make a good poker agent going. Even without deep learning, I still have cards worth playing left.

At a certain point I will be done with Spiral, done with CFR agents, done with graphical interfaces. The time will come to get back into ML. I will start with probabilistic programming and go through all the math in a dependently typed language.

I will revisit Coq and instead of trying out tactics blindly, I will investigate what those rewrites are doing.

I'll master formal math and go at it as a mathematician. Hopefully by that point all the programming training will act as a springboard into it.

It is fine.

Deep learning is a short cut and shortcuts are longshots anyway. That exploration idea that I had is great, but it won't solve the roboustness problems that I am having with DL on toy tasks.

Once I have the correct principles - then i will win. It will be done through the power of reasoning, not through hail mary passes.

I'll accumulate my power in slow tedious steps.

It has been 4 years since I started. I wonder what sort of ability will I have after 10?