Done with resting. Let me finally resume. This was such an extended break. Actually, is it even that longer than usual? 2.5 hours is not that long.
Let me just focus on the error messages for now and then I will deal with the testing harness.

Thought I've nearly finished the compiler the testing will be a decent chunk of work after that.

Damn, I am such a harder worker even though I do not want to be. It is like a punishment for my grandiose ambitions. I really wish somebody else did Spiral back in the 90s or 00s.

Well, once I do this, the thing will be done for good. Just two more weeks of this.

The eventual Pharo port will be continuation of this, but it won't be a wholesale redesign. It will be a lot more straightforward apart from the IDE parts. It won't be nearly as mentally exhausting as this is.

For one, I won't have to redo all the libraries all over again. The language featureset should be stable past this point.

...There is nothing more to day. Let me do my deed for the day.