"8:25am. I am up. Let me just say one thing - The Gamer is really, really boring. I should have dropped it 100 chapters ago. It feels like half the chapters are some useless lore dump and the other half is people getting astonished that Gods, Demons, Monsters and such exist. 'Abyss is a bad place, m'kay?'
It is really a pity since the early chapter were really great.

8:30am. Before I start let me talk a little about the design of the libraries.

Now that I have objects, I am agonizing whether I should use `Array.map f x` like in F# or `x (map: f)` like in Pharo.

One thing I am considering is getting rid of partial application where it is not necessary such as in folds.

While I wavering on the issue of OO vs functional design, there is something I understand. There is something about keyword arguments that immediately jumped out at me in Pharo. Even as a complete newcomer to the language, I was really impressed how readable it is.

When writing a function it does not really matter if it is something like...

```
inl access key value = ...
```

or

```
inl access (key, value) = ...
```

or

```
inl access {key value} = ...
inl access key:value: = ...
```

What really does matter is what happens months later when you trying to use the thing. At that point you've forgotten which one of the four is necessary. Is it `key value` or `key, value` or even `value key`? For `array_create` I've frequently forgotten whether the type or the size should come first.

And when you read code and see something like...

```
access a b
```

The only way to know what `a` and `b` are is to be familiar with `access`.

Suppose the function was named poorly, something like:

```
qwe a b
```

Now you have absolutely no idea what the thing could possibly be and the issue is that much worse with poor IDE support. Now with keyword arguments, the library author and the user is forced to be explicit.

```
qwe (key: a value: b)
```

There is an immediate vast improvement in readability. And this factor is independent of whether I am using OO or functional design.

8:50am. Hmmm, let me chill a little.

If I am going to be doing functional design with some OO sprinkled in, I should really figure out what my principles should be. What would be the most comfortable way of using the objects to facilitate such a style?"