As hard as it is to believe, I think I've done enough work to finally run Spiral v0.1 for the first time. I have a really, really long debugging session ahead of me now.
...

Yeah, it will be a pain. It will be a pain.

All the type errors and the refactoring and the debugging - it will be a pain. The Pharo segment of my life after this is done will definitely be a lot more interesting.

The accumulate skill will eventually have to point me in some useful direction. I will continue to believe in that.

Up to then...

It is time to do some real work.

This will be last time I do 'work' of this sort. Programming should ideally be a lot more like play.

After I am done with the basic tests, I won't go mad trying to rewrite the libraries since I know I will have another redesign when the time comes to do Spiral yet again in Pharo.

I'll let need drive my rewriting efforts. Honestly redoing everything in Spiral just to test the exploration method is just too much work when I know the method won't do better in a way that matters.

The ML library is pretty deep on the abstraction scale and uses all the libraries. Trying to get it back online would add at least two and more likely 3 weeks on top of my schedule.

...

Well, let me run the first test.