"2:30pm. Done with chores.
Now is the finally time for `test1`.

```
let test1: SpiralModule =
    {
    name="test1"
    prerequisites=[]
    description="Does it run?"
    code=
    """
inl a = 5
inl b = 10
a + b
    """
    }
```

The test is only this, but parsing, prepass, peval, the `Core` library. There is a lot of that can go wrong in between. It won't be that easy to get this to work.

I have absolutely no idea what the amount of bugs I encounter will be, but there will be some no doubt.

2:35pm. I can only take courage and dive in. It should not be too bad.

Let me give it a shot."