"2:45pm. Maybe it has something to do with all those globals I am using now?
v0.09 works, but v0.1 does not.

2:50pm. https://stackoverflow.com/questions/42665926/cannot-resolve-dependency-to-assembly-fsharp-core-4-4-1-0-when-using-vs-2017/42666073

Maybe I could try downloading F# Core of Nuget and see if that does anything?

2:55pm.

```
printfn "Hello"
```

Even with a single file with just this, I still get the same error.

Aggggghhhh, I spent 20m on this already. This is such bullshit.

I just noticed that the project does not have a reference to `Fsharp.Core`.

Well, let me take this chance to upgrade to 4.61.

3pm. Ok, I finally manged to get through that retarded error.

Let me finally run `test1` and get the real testing under way."