Right now I am just chilling, but I'll start the work soon. I should not have wrecked the prepass yesterday. I was tired and I jumped the gun a bit on what the error should be.
Well, let me finish those Gridman yuri doujins that were at the top of Mangadex when I checked and then I will start the day.

Ever since summer, I had not touched games as keeping myself from immersive activities other than programming is important for maintaining motivation.

Plus it was kind of a deal that I had with myself until I made an achievement with RL. Games are for the agents.

I think once I move to the next stage my luck will improve.

All languages are a manifestation of a reasoning pattern and with Spiral, all I can really do is make an optimal implementation of the ideal algorithm. This would be great if I knew what that the ideal algorithm is, but since I do not I am in the situation I am now.

There is a weak link between being able to make algorithms fast and making the right algorithms in the first place.

The interfaces I am going to make starting from here will be different. Being able to do these things will have the strongest connection possible with algorithmic search.

If I become able to make the right kinds of 'visualizations' for the lack of a better word, that would mean that I understand the algorithms I am trying to tame completely. It will act as a beacon in the darkness of the design space. It will guide me to where I need to go.

Ultimately, the things I am looking to guide the learning process reflect my understanding of it. If I am just looking at the cost then that is my level. I think those capable of causing the Singularity will have a much deeper insight with regards to what needs to be kept track of.

Focusing on interfaces is certainly a better idea than churning out all those worthless papers that the research community is currently doing.

I will turn this around.

The reasoning pattern I've developed in Spiral is far from worthless. I just need to evolve it into the next form. Eventually the code I am writing will become an asset.

This next part of my journey will definitely be more fun than the last.