"1:35pm.
```
    let char_quoted s = (between (skipChar '\'') (skipChar '"' >>. spaces) (char_quoted_read (fun _ -> true)) |>> LitChar) s
    let string_quoted (s: CharStream<_>) = (between (skipChar '"') (skipChar '"' >>. spaces) (manyChars (char_quoted_read ((<>) '"'))) |>> LitString) s
```

I needed to put spaces at the end.

Damn that thing.

```
let string_raw = string_raw_template "@\""
```

This is also broken.

```

    let inline string_raw_template str str' =
        skipString str >>. charsTillString str' true Int32.MaxValue .>> spaces
        |>> LitString

    let string_raw = string_raw_template "@\"" "\""
    let string_raw_triple = string_raw_template "\"\"\"" "\"\"\""
```

Even after I've done this it still fails. How incredibly difficult these parser errors are. They remind me of Spiral's type errors.

1:45pm. I admit I am losing my spirit a bit. Damn the break. I had good momentum before I started to get introspective about the state of the parser.

For the sake of speeding up the parsing literally nothing I could ever do could be as helpful as tokenizing.

It would also really help me in figuring out what the parsers are doing wrong.

1:50pm. Focus me. Let me do the raw string template.

1:55pm.

```
        let case_parser_macro =
            inbuilt_op_core '@' >>= fun a _ ->
```

Ah, I guess this explains why I previously used `operatorChar` in `inbuilt_op_core`.

...Look me.

The reason why I am having so much trouble here is because of conflict of interest. One one hand I want a fast parser. On the other hand I want to do better than last time - I do not want to sprinkle the entire codebase with `attempt`s.

I want lookahead to be as easy as possible.

2pm. I wrecked the entire parser this time for essentially no reason.

2:05pm. Maybe I wanted to motivate myself...

2:15pm. Ok, I finally fixed it. Even though I am now depressed because of the parser, let me continue going forward. I will get through this.

```
inl not x = !EQ(x,false)
          ^
Expecting: ! or #, identifier, keyword, number, '"', '"""', '&', '\'', '(', ','
, '->', '.', '//', ':', '::', '=>', '@"', 'as', 'false', 'true', 'when', '{' or
'|'
```

Just what could this possibly be?

```
    let f (s: CharStream<_>) =
        printfn "in f"
        printfn "%c%c%c" (s.Peek(0)) (s.Peek(1)) (s.Peek(2))
        Reply(())
```

Let me save this thing here. I am noticing that I keep rewriting it.

```
in f
!EQ
Error in test1: Ln: 2 Col: 11
inl not x = !EQ(x,false)
          ^
Expecting: ! or #, identifier, keyword, number, '"', '"""', '&', '\'', '(', ','
, '->', '.', '//', ':', '::', '=>', '@"', 'as', 'false', 'true', 'when', '{' or
'|'
```

Ok, so this is what I would expect.

2:30pm.

```
inl not x = !EQ(x, falser)
```

It only really fails when I put in `false` or `true`. When it is like this or any other variable it sort of works...

It tells me that the variable is not bound which is an indication that the stupid thing is still not compiling correctly.

To say that I am having extreme difficulty with these crazy parser errors would be an understatement.

I am completely lost at the moment.

Do I really want to continue plowing on like this or do I want to try a different tack?

Redoing the parser from the ground up is not a trivial undertaking. Sure I will be able to reuse stuff from the current one, but...

2:40pm. Yeah, I am better than this.

I am going to tokenize. I'll stop rellying on FParsec so much. Just because it has a feature does not mean I have to use it.

Parsing is no trivial task by any means. It is definitely harder then the other passes.

Hence I must make it easier.

This actually has a lot of similarity with deep learning.

It always comes down to how to propagate information through the graph in an optimal manner. Just like there are optimizations for performance, there are optimizations for composability.

To optimize is to understand.

3:10pm. Right now I am thinking of how I would do the tokenization.

I am especially concerned with some of the things like that fucking negation operator which requires some crazy rules to deal with.

You know what I will do regarding that negation operator is only have the special rule at the start of an expression. Similarly like for the `(+)` or `(-)` unapplied operators.

Meaning I'll allow `inl x = -2` or `inl y = -x` or `(-x)`, but `inl x = y - 2` will be the same as `inl x = y -2`.

There is something I could try though...

`inl y = 2 * 4+10 ` ...

Well, I'll leave that as `inl y = 2 * 4 + 10`.

But I will allow tight binding for `inl y = f x.z` or `inl y = f [w = 3].w`

3:25pm. Ok, I've decided. I'll go for tokenization. I'll split parsing into stages that can be more easily digested.

3:30pm. I will do what I should have done long ago.

This is my punishment for not doing the right thing the first time around.

I am just so depressed right now. Those parser errors earlier pretty much killed my momentum. Errors are fine, but errors that are too hard to deal with are a problem.

I will attain the pinnacle of simplicity in the parser. By the time I am done I will not have to say that the parser is Spiral's most complicated part.

3:35pm. Let me start by backing up the parser.

Then I'll do it from scratch."