"3:50pm.
```
type TokenPosition = {
    start: int
    end_: int
    line: int
    }

type TokenSpecial =
    | SpecMatch
    | SpecFunction
    | SpecWith
    | SpecWithout
    | SpecAs
    | SpecWhen
    | SpecInl
    | SpecInm
    | SpecInb
    | SpecRec
    | SpecIf
    | SpecThen
    | SpecElif
    | SpecElse
    | SpecTrue
    | SpecFalse
    | SpecOpen
    | SpecJoin
    | SpecType
    | SpecTypeCatch

type SpiralToken =
    | TokVar of TokenPosition * string
    | TokMessageUnary of TokenPosition * string
    | TokMessage of TokenPosition * string
    | TokNumber of TokenPosition * string
    | TokOperator of TokenPosition * string
    | TokAnnotatedOne of TokenPosition * SpiralToken // ! Used for the active pattern and inbuilt ops.
    | TokAnnotatedTwo of TokenPosition * SpiralToken // @ Used for parser macros
    | TokAnnotatedThree of TokenPosition * SpiralToken // # Used for the unboxing pattern.
    | TokAnnotatedFour of TokenPosition * SpiralToken // $ Used for the injection in patterns and in RecordWith
    | TokSpecial of TokenPosition * TokenSpecial
```

I am not sure that when all is said and done this will be a memory saving. Well, I'll hash cons the operator tokens at least.

For operators I should also put in the associativity and precedence.

3:55pm. Oh yes, cannot forget the brackets.

4pm.

```
type SpiralToken =
    | TokVar of TokenPosition * string
    | TokMessageUnary of TokenPosition * string
    | TokMessage of TokenPosition * string
    | TokNumber of TokenPosition * string
    | TokOperator of TokenPosition * TokenOperator
    | TokBracketRound of TokenPosition * SpiralToken list
    | TokBracketCurly of TokenPosition * SpiralToken list
    | TokBracketSquare of TokenPosition * SpiralToken list
    | TokComma of TokenPosition
    | TokSemicolon of TokenPosition
    | TokAnnotatedOne of TokenPosition * SpiralToken // ! Used for the active pattern and inbuilt ops.
    | TokAnnotatedTwo of TokenPosition * SpiralToken // @ Used for parser macros
    | TokAnnotatedThree of TokenPosition * SpiralToken // # Used for the unboxing pattern.
    | TokAnnotatedFour of TokenPosition * SpiralToken // $ Used for the injection in patterns and in RecordWith
    | TokSpecial of TokenPosition * TokenSpecial
```

I am realizing that I am going to have a hard time if I want to extend the language syntactically past this point.

The keyboard is kind of tapped of operators by this time.

It is a good thing then that the language has everything it needs.

I will do this thing.

...Ah, there is something I should add.

```
| SpecWildcard
```

Now I am set.

Let me make the tokenization functions.

A lot of the difficulty in the parser is actually dealing with the conflicts that arise from similarly parsed terms. Having the differences disambiguated at the tokenization stage will make the rest flow a lot easier.

For the rest of the day, let me just focus on the tokenizer and then I will stop. Tomorrow I will catch wind again and deal with the parser properly.

And doing this like so will force me to put proper error messages during parsing."