"9:55am. Let me resume. It is time for finish the tokenizer.
10am. Ah, yes, can't forget spaces. They need to skip comments.

```
let bracket x s =
    let start = pos s
    between (skipChar x >>. spaces) (skipChar x >>. spaces)
    //| TokBracketRound of TokenPosition * SpiralToken list
    //| TokBracketCurly of TokenPosition * SpiralToken list
    //| TokBracketSquare of TokenPosition * SpiralToken list
```

I'll do these last. I forgot that I need the complete tokenizer for this. Let me do the simpler cases first.

```
let annot_one s =
    let start = pos s

    //| TokAnnotatedOne of TokenPosition * SpiralToken // ! Used for the active pattern and inbuilt ops.
    //| TokAnnotatedTwo of TokenPosition * SpiralToken // @ Used for parser macros
    //| TokAnnotatedThree of TokenPosition * SpiralToken // # Used for the unboxing pattern.
    //| TokAnnotatedFour of TokenPosition * SpiralToken // $ Used for the injection in patterns and in RecordWith
```

This needs recursion as well.

One thing I've forgotten to do are string literals. Let me take care of that now.

10:25am. Ok...

10:40am.

```
let rec tokenize s =
    let inline bracket f l r s =
        let start = pos s
        (
        between (skipChar l >>. spaces) (skipChar r >>. spaces) (many tokenize)
        >>= (fun x s -> Reply(f (tok start (pos s), x)))
        .>> spaces
        ) s

    let bracket_round s = bracket TokBracketRound '(' ')' s
    let bracket_curly s = bracket TokBracketCurly '{' '}' s
    let bracket_square s = bracket TokBracketSquare '[' ']' s

    let unary s =
        let start = pos s
        match s.Peek() with
        | '!' -> s.Skip(); (tokenize |>> (fun x -> TokUnaryOne(tok start (pos s), x)) .>> spaces) s
        | '@' -> s.Skip(); (tokenize |>> (fun x -> TokUnaryTwo(tok start (pos s), x)) .>> spaces) s
        | '#' -> s.Skip(); (tokenize |>> (fun x -> TokUnaryThree(tok start (pos s), x)) .>> spaces) s
        | '$' -> s.Skip(); (tokenize |>> (fun x -> TokUnaryFour(tok start (pos s), x)) .>> spaces) s
        | _ -> Reply(Error, expected "`!`,`@`,`#` or `$`")

    choice
        [|
        var; message_unary; number
        string_raw; string_raw_triple; char_quoted; string_quoted
        bracket_round; bracket_curly; bracket_square; unary
        comma; semicolon; operator
        |]
        s
```

Here is `tokenize`. I did a lot of cleaning up besides this. Rather than call them annotations, I decided to make `!`...`$` the unary operators. This is how they should be. None of that endless lookup that I cannot decide what it should be. Whether they are unary or not is decided during tokenization.

The same for the brackets. I think doing this at this stage is a good decision.

10:55am. Did some filling in. With this the tokenizer is complete. These 300 lines will make parsing a lot simpler.

No more thinking about backtracking. I will do it all in one blow.

11am. There is something I will do. I will rename `TyKeyword` into `TyMessage` and such. Having them as keywords makes things confusing.

I'll do that when I am done with all of this. I wonder if the IDE has a automatic rename in it somewhere? It probably does, I just never used it.

...Ah, yes it does. Ok.

I'll forget this for the time being since it won't be a problem.

11:05am. The next thing in line is the parser.

But before that let me take a break. Once I am done with the parser, the rest will go by a lot easier.

This parser, this damn parser. It is nothing, but trouble. It has always been more complex than it should have and I am glad that I will finally have the chance to tame the beast. This has been a long thing coming."