I need to start work on the parser. Uggghhh...
But really the parser need not be that difficult. I am just tired of it. Parsing of programming languages is not a subject I have a particular interest in.

```
    let rec expr s =
        let expressions s = type_union ^<| keyword_message ^<| tuple ^<| operators ^<| application ^<| immediate_keyword_unary ^<| expressions expr <| s
        let statements s = statements expr <| s
        annotations ^<| indentations statements expressions <| s
```

This is what I essentially have to implement. Once I am done with this, I will be done with the parser.

I can reuse the structure, but not the code for this unfortunately. This is why I did not want to go for tokenization last time.

Well, it is fine. I won't rush myself to finish this. Let me go at it step by step. I'll allocate a few days for dealing with this.

This really just had to jump out at me during testing and scramble my plans.

Well, it happens. Just think of how great Spiral will be once this is done. That will keep me focused on the goal.