"4:25pm. Finally done with chores. Another chapter of that delinquent magical girl manga came out a few hours ago.
https://mangadex.org/title/20573/machigatta-ko-wo-mahou-shoujo-ni-shite-shimatta
https://mangadex.org/title/34522/himenospia

And I missed two chapters of Himenospia. It seems there wasn't a thread on /a/.

4:45pm. I am just sitting here and thinking about how am I going to deal with the errors. I made a serious mistake by putting ints as the other field of the error instead of the token.

This is something I am going to fix.

I'll deal with the tokens somehow. This won't be hard.

There is one other thing I made a mistake about. I should have stuck to the FParsec semantics for the `<|>` and `choice`.

Literally the only place that needs backtracking is the module parser.

Right now I am really unsure of what the ideal way to provide good error messages for that thing would be. I am at a loss.

Rather how do I do the initial parse smoothly?

5pm. I really want to read manga, but I just can't stop thinking about this. I think I understand how.

FParsec was right after all. I'll revert `<|>` and friends to their previous semantics.

Here is what I have to do.

`(attempt (module_with_prefix .>> (with <|> without)) .>>. module_body) <|> module_body`.

`module_with_prefix .>> (with <|> without)`

The key part of figuring out how to do good error messages here is deciding where to backtrack. If I backtrack all the way by wrapping the entire thing in an attempt that will produce poor results.

5:05pm. Hmmmm...

I need to go a bit further in my reasoning though.

The problem with the above scheme is that `with` and `without` can come anywhere in the module.

Maybe I could change that syntax...

Let me take a look at the libraries. How many times have I used `without` together with `with`?

5:10pm. Why is find so much slower when it is case sensitive? Shitty IDE.

5:15pm. Ok, there were uses. I know how I am going to do this now. I'll combine the pattern I have now together with the above.

Let me take care of the tokens and this and I will call it a day. I'll finish the parser properly tomorrow and get back to testing.

5:30pm.

```
    member inline d.TryCurrentTemplate on_succ on_fail =
        let i = d.Index
        if i > 0 && i < d.Length then on_succ d.l.[i]
        else on_fail()

    member inline d.TryCurrent f = d.TryCurrentTemplate f (fun () -> Fail [])
    member d.FailWith er = d.TryCurrent (fun x -> Fail [x,er])

    member inline d.LineTemplate f = d.TryCurrentTemplate f (fun _ -> -1)

    member d.Module = d.module_
    member d.Line = d.LineTemplate (fun x -> x.Start.line)
    member d.Col = d.LineTemplate (fun x -> x.Start.column)
    member d.LineEnd = d.LineTemplate (fun x -> x.End.line)
    member d.ColEnd = d.LineTemplate (fun x -> x.End.column)

    member inline d.Skip'(i) = d.i := d.i.contents+i
    member d.Skip = d.Skip'(1)
```

Yeah, this refactoring is good. Accessing the current token should all be done from a central location.

I want to eliminate all possibility of accessing out of bounds elements.

5:45pm.

```
let inline attempt a (d: ParserEnv) =
    let s = d.Index
    match a d with
    | Ok x -> Ok x
    | Fail a as a' -> d.IndexSet s; a'
```

6pm.

```
        let record_with =
            let withs' s = many1 record_create_with s
            let withouts' s = many1 record_create_without s
            let withs s = (with_ >>. withs') s
            let withouts s = (without >>. withouts') s
            attempt
                (tuple3
                    ((var |>> v) <|> rounds expr)
                    (many ((keyword_unary |>> Types.keyword_unary) <|> (dot >>. rounds expr)))
                    ((with_ >>% withs') <|> (without >>% withouts')))
            >>= (fun (init,names,next) s ->
                pipe2 next (many (withs <|> withouts) |>> List.concat) (fun a b -> mp_with init names (List.append a b)) s
                )

        let record_create = many record_create_with |>> mp_create

        curlies (record_with <|> record_create)
```

Maybe I'd really have been better off trying to parse `record_create` first.

```
        let record_create_with s = (parse_binding_with .>> optional semicolon) s
        let record_create_without s = (parse_binding_without .>> optional semicolon) s
```

I was not using the right semicolon parser here previously.

6:25pm. Done with lunch.

```
pipe2 next (many (withs <|> withouts)) (fun a b -> mp_with init names (List.concat(a::b))) s
```

I think I will go for this actually. There are other orders which I could attempt, but this will suffice.

7:10pm. I've stopped a while ago so let me commit here. I am dead tired and want to watch Kotobuki.

The parser is over 90% done at this point, I just need to put the finishing touches on how errors are being handled and clean things up. After that I will be able to get back to testing.

Now that I've cleaned up the parser I do not think it will be giving me much trouble anymore. I can expect bugs, but they won't be nearly as severe as those that greeted me a few days ago. I will be able to go through them soon enough.

The same goes for the partial evaluator and the codegen. The entire language is super straightforward now despite being long and there won't be any issues. I'll be done with the tests by the end of the month.

Then the things will get real. It is not the kind of step I wanted, but it will be the one I need to make.

This current machine learning wave sucks. I had imagined that I would get where I want by waiting for the field to get to me, but I think I'll just get there with my own power instead, in a way only I can.

I have everything I need for systems programming in Spiral. The next step would be to master virtual machines and higher level programming.

I deeply regret that all the code that I wrote in 2016 in F# and in v0.09 of Spiral is just going to be abandoned. Had I really done a good job I could have made that my power. It is true that the work that I did will serve as the template for what is to come, but it is like a grazing blow.

The next one I send out I will make sure that it sinks deeply. I won't let it end so easily.

As long as I have life, I will continue fighting."